chore: add auth helpers and health check
